Refactoring: remove old OC Collector code
Initial phase of removal of old OC code that is still present on the repo.
