+++ application/builder/receivers_builder_test.go
