+++ application/builder/testdata/global_attributes_all.yaml
