+++ application/builder/testdata/global_attributes_key_mapping.yaml
