+++ application/builder/testdata/receivers_disabled.yaml
