+++ application/builder/testdata/receivers_enabled.yaml
