+++ application/logger.go
@@ -12,7 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package sender contains specialized senders to different backends. Unlike
-// exporters they do not buffer or attempt to resend that failed batches, all
-// of that is delegated to the users of the senders.
-package sender
+package application
+
+import (
+	"flag"
+
+	"github.com/spf13/viper"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+)
+
+const (
+	logLevelCfg = "log-level"
+)
+
+func loggerFlags(flags *flag.FlagSet) {
+	flags.String(logLevelCfg, "INFO", "Output level of logs (TRACE, DEBUG, INFO, WARN, ERROR, FATAL)")
+}
+
+func newLogger(v *viper.Viper) (*zap.Logger, error) {
+	var level zapcore.Level
+	err := (&level).UnmarshalText([]byte(v.GetString(logLevelCfg)))
+	if err != nil {
+		return nil, err
+	}
+	conf := zap.NewProductionConfig()
+	conf.Level.SetLevel(level)
+	return conf.Build()
+}
