+++ application/testdata/otelsvc-config.yaml
