--- cmd/occollector/Dockerfile
@@ -1,8 +1,0 @@
-FROM alpine:latest as certs
-RUN apk --update add ca-certificates
-
-FROM scratch
-COPY --from=certs /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
-COPY occollector_linux /
-ENTRYPOINT ["/occollector_linux"]
-EXPOSE 55678
