+++ exporter/jaegerexporter/jaeger_proto_grpc_sender.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sender
+package jaegerexporter
 
 import (
 	"context"
