--- internal/collector/zipkin/.nocover
@@ -1,1 +1,0 @@
-FIXME: Move this to the receiver.
