--- internal/collector/zipkin/scribe/.nocover
@@ -1,1 +1,0 @@
-FIXME: Move this to the receiver.
